# Lab7/Part2/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Product API",
    "description": "API for managing products",
    "version": "1.0.0"
  },
  "paths": {
    "/api/products": {
      "get": {
        "summary": "Get a list of products",
        "tags": ["Products"],
        "responses": {
          "200": {
            "description": "Successful response with a list of products",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "title": "Product 1",
                    "description": "Description of Product 1",
                    "price": 19.99,
                    "image": "https://example.com/product1.jpg",
                    "category": "Electronics"
                  },
                  {
                    "id": 2,
                    "title": "Product 2",
                    "description": "Description of Product 2",
                    "price": 29.99,
                    "image": "https://example.com/product2.jpg",
                    "category": "Clothing"
                  }
                ]
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": { "message": "Internal server error" }
              }
            }
          }
        }
      }
    },
    "/api/products/{category}": {
      "get": {
        "summary": "Get products by category",
        "tags": ["Products"],
        "parameters": [
          {
            "in": "path",
            "name": "category",
            "required": true,
            "description": "The category of products to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with products of the specified category",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 3,
                    "title": "Product 3",
                    "description": "Description of Product 3",
                    "price": 39.99,
                    "image": "https://example.com/product3.jpg",
                    "category": "Electronics"
                  }
                ]
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": { "message": "Internal server error" }
              }
            }
          }
        }
      }
    }
  }
}